Remove all Clang warnings from core library (#1666)
Description
When compiling Lethe with Clang, multiple warning arise throughout the code. We aim to remove these warnings so that Lethe compiles on both GCC and Clang without any warnings. This changes removes all the warning in the core library. This PR addresses #1648 .

Testing
None of the tests should be affected.

Documentation
Nothing to document here