Add mesh translation and rotation functions (#1668)
Description
The parameters for initial translation and rotation of the mesh were included in the Parameters::Mesh structure, but not used in attach_grid_to_triangulation. This PR fixes this.

Solution
The functions GridTools::shift and GridTools::rotate are now called using the parameters translation, rotation_axis, and rotation_angle.

Testing
A new application test for a 3D MMS has been added.