Fix issue by attaching dof_handler to DataOut instance + refactor structs to include cell data (#1661)
Description
The output_field_hook functions were not working due to the last refactoring of the write output results, PR #1624. The reason is that we were not attaching a default DoF handler to the DataOut instance. Additionally, the inclusion of cell fields, mainly the subdomain field, was done without the struct.

Solution
Call attach_dof_handler upon instantiating DataOut.
Refactor output_field_hook function to use the new struct.
Add a variant of the struct for the case of cell data.

Testing
This was first observed when @blaisb was running a sharp example. I tested the "Sedimentation of One Particle" sharp example, and it worked just fine without the hotfix of PR #1658.

Documentation
No need. Just a fix.

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>