Work in progress, but still not ok.
So the idea of projecting and everything is implemented, but the current
implementation is not working at all. All the ingredients are there,
but the particle-force field always remains zero for a reason I ignore.
I will move on to writing a unit test for that instead of fighting this
head-on