Add missing include for matrix-free internal function. Important for deal.II 9.8-pre. (#1676)
Description
The new refactoring of deal.II matrix-free prevented the compilation of Lethe due to a missing include.

Solution
A missing include "deal.II/multigrid/mg_transfer_matrix_free.templates.h" was added to enable compilation on deal.II master version (9.8-pre)

Testing
Compilation works fine on my machine and on the CI now. There are some warnings to fix in the future and we will need to wait until the deal.II docker image is updated for the broken test to fix itself.