Fix the include guards in the matrix-free solver (#1677)
Description
A mistake was made in the previous (by me) in which the include guards were not correct for deal.II 9.7 when accounting for the new matrix free include. This was rapidly found by @peterrum

Solution
Fix the include guard to have the correct deal.II version and add the else clause.

Testing
Compiles on my machine for both master and 9.7