Only cancel CI when is on a Pull request (#1675)
Description
We rely extensively on the CI, not only to check if the code works before merging, but also when we do git bissect or other features. One of the main thing that we added in the last few years is CI cancellation when there is another PUSH on top of the same branch, this prevents the CI from clogging because of a ton of commit push when people are not very careful. This behavior is nice, but it is not desirable on the master branch because it can give the impression that the CI crashed (and will give a red check) while the CI would actually have been good. This happens a lot when we merge two branches to master within the spawn of an hour.

Solution
I removed the cancel-in-progress when the branch is not a pull request. This is sufficient to fix the issue.