Added AssertThrow in the read_checkpoint function in the DEM (#1687)
Description
If the checkpoint controller does not exist while reading a restart at the beginning of a simulation, no error message was being written and the simulation was crashing. This was annoying when you are copy-pasting a .prm file that already exist, change some parametesr manually and then start your new simulation while forgetting that the restart parameter is still at true. (Aka, me losing 10 minutes this morning wondering why my prm was not working)

Now, with the new AssertThrow, the user is informed that the checkpoint_controller is not found, thus he know that we are trying to restart a simulation. If he doesn't want to restart, he knows right away what to do. On the other hand, if he wants to restart, well he knows there is a problem.