Table serialization (#1673)
Description
This PR addresses project task #1159. After the changes, the tables are now serialized and deserialized in NavierStokesBase, CFDDEMSovler and FluidDynamicsNietsche. Once a checkpoint is written, the TableHandler objects of NavierStokesBase are written, in addition to those particular to the solver.

Testing
This has been tested by checkpointing and restarting simulations involving the three solvers mentioned. The tables were printed at the restarting point of a simulation, and contained the data from iterations previously run.

Documentation
No need for documentation