Mf vans cfdem coupling (#1686)
Description
This pull request introduces a first version of the following components:

Particle-fluid force projection onto the CFD mesh
Matrix-free VANS solver with particle-fluid coupling.
The current coupling between particles and fluid is fully explicit, meaning that there is no mass matrix associated with the drag term. This is clearly not very stable for fluids which are not dense. Furthermore, the drag formulation is not complete since there is no division by density of the fluid (since density of the fluid is not a supported physical property at the moment in the VANS solver).

Consequently, the coupling should be treated as highly experimental. I am mostly opening this PR right now to ensure that we can checkpoint at this current stage since @antonellasuccar will need the features for the unit tests. After that, the rest of the work to be done for this coupling is on my todo list and should go smoothly in another PR.

It includes:

Considering the fluid density in the drag force (divided the drag force by fluid density at the quadrature point)
Establishing an implicit coupling for the drag force by projecting it differently than the sum of the forces
Developing a CFD-DEM coupling engine with the matrix-free. This will be a combination of the vans-matrix-free solver and the cfd_dem_coupling one. I will try to avoid as much as possible the usage of copy-pasting.
Validating and benchmarking the final solver.
I estimate this will be a bare minimum of 3 other PRs and I'd rather take the time to do this smoothly.

Testing
This PR also introduces the first batch of unit testing for the particle-projector. There are two unit tests that are into place.

Tests the calculation of an homogenous void fraction onto a unit cube when there are a large amount of particles (32k)
Tests the calculation of the void fraction AND of two different particle force profiles.
This opens up the way for more unit testing of the projector (e.g. velocity, etc.). This is something I will be continuing to do! These new unit tests are very interesting in that respect.

Furthermore, I have added an application test at the moment which is simply a flow through the packed bed, but where there is coupling between the fluid and the particles through drag.

Documentation
The new solver should not be documented thus far since it has no new parameters and is absolutely not production ready.
Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>
Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>