Gaussian reaction model (#1688)
Description
In the tracer physics, the reaction models for heterogeneous catalysis using lethe-fluid-sharp only included the tanh model, which at best could be used to impose a reaction term inside the solid while setting it to 0 in the fluid. To be able to set a reaction term only at the interface, to mimic the effect of active sites only at the surface of the solid, a Gaussian model is implemented instead. A similar model has been added to the tracer diffusivity, eventually to hinder exchange between phases, for example.

Testing
A new test has been added:

applications_tests/lethe-fluid-sharp/mms_tracer_reaction_heterogeneous_gaussian
Documentation
The new model has been added to the documentation of the physical properties.

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>