Cleanup periodic parameter names with an "_" (#1698)
Description
We have some parameters in lethe which use _ in their declaration instead of spaces. We do this for very few parameters and it is done in an arbitrary and illogical way (mostly me creating these parameters like 5 years ago). This PR refactors this by removing some of the _ in the parameters, namely the periodic ones. I have declared an alias for the time being to not break people's prm file, but this will be deprecated anyway.

This PR partially adresses #1647

Testing
No test should change.

Documentation
Documentation has been refactored to reflect the change.