Clean up examples CI + release CI limit mpi cores (#1699)
Description
The release CI uses a regular expression to exclude the mpirun=2 tests from the first run but does not do the same for the mpirun=3 tests.

Solution
This PR addresses this by instead making all mpirun=1 test run first and then running the mpirun=2 or 3 tests afterwards. This should be more robust and also faster

Testing
None of the tests are changed, it should just make the release CI more robust.

Documentation
Nothing to document here.