Boundary condition id as list (#1697)
Description
One major request of users in the last few years have been to be able to give multiple boundary id when specifying a boundary condition. ( see for example #1331 ).
This PR achieves this by parsing a list for the boundary condition and duplicating the boundary condition using all of the ids in the list. This is a simple yet efficient mechanism to enable list for boundary conditions.

I have taken the opportunity to add additional defensive assertions to ensure that the boundary ids are not negative and so on and so forth. This feature is now supported by ALL physics.

Testing
I have modified an existing test (the cavity one) to use the list specification instead of duplicating the boundary condition subsection. I have also modified the cavity example to use this specification also. I have taken the opportunity to modify one of the tolerances of the cavity example to ensure that the results are sufficiently converged.

Documentation
I have updated one of the new example to use the new feature, but I don't think this requires specific documentation?

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>
Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>
Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>