Add pre-commit hook for copyright (#1700)
Description
I find it very annoying to manually update the copyright years. Consequently, I have added a script that automatically update the copyright year. It is a bit (a lot) hackish and uses a blend of bash and perl, but it does the trick. I would suggest using it with some care for now.

Testing
I have tested the script on a few use case and it works well it seems.

Documentation
I don't think this needs to be documented since this is in the pre-commit hooks.

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>