Fix an issue with the timer requesting (#1705)
Description
The timer in navier_stokes_base post_process_fd was requesting the time summary while the timer itself was running. In the recent deal.II change, this undefined behavior is asserted for.

Solution
This commit manually opens and closes the timing at the right locations to ensure that the timer is not running anymore when the time is requested.

Testing
All current tests should remain unchanged

Documentation
Nothing to document