Restore CLANG TIDY (#1708)
Description
The Clang TIDY is back. This CI step is there to ensure the code is as tidy as possible. Currently the Clang tidy workflow still uses clang tidy 12, but i will move it to a newer version in a follow-up PR.

Testing
A few minor mistakes were found mostly in the tests and related to arrange based loops or passing by const reference instead of copy. I have taken the liberty of fixing these issues at the same time!

Documentation
None required.