Bug fix - Allow CFD-DEM simulations to use an adaptive time-step. (#1715)
Description
The CFD-DEM implementation in Lethe had a bug when it was used in tandem with adaptive time-stepping. The DEM time-step was never recalculated after the start of the simulation, which meant that although the CFD time step was changing, the DEM time-step was not changing as the simulation progressed. This is not an issue for the examples or the previous work we made, since the CFD-DEM was never tested or validated with adaptive time-stepping in mind. However, this is still a bug.

Solution
I have added a new simulation_control function which gives back if the simulation uses adaptive time-stepping or not. If a simulation uses adaptive time-stepping, the DEM time-step is recalculated when the coupling is done. This is a correct fix for now, but in the future I will refactor it so that the SimulationControl in the CFD-DEM simulation is actually a specialized version that has both the CFD and the DEM time-step. This is not necessary for now, but I am putting it in my todo and opening an issue for it.

Testing
I have added a new test case which is the single particle sedimentation with a CFD time step which is adaptive.