Refactor DEM to be its own section in the theory guide (#1713)
Description
@OGaboriault has always been slightly annoyed that the DEM section for granular flows only lived in the CFD-DEM section. I now agree with that notion.

Furthermore, I felt it was a bit weird the way we subdivided thing. In the multiphysics section, we refer to things are physics, whereas in the multiphase section we referred them as numerical methods. I have refactored this to make more sense (in my opinion). Now essentially we describe the type of multiphase flows and present the numerical method. I know DEM is not necessarily Multiphase, but you could say particles are always a bit of a multiphase flow even though we don't model the void and I feel this is a bit clearer.


Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>