Fix vof initial condition geometric smoothing (#1718)
Description
There was a seg fault when the geometric smoother was selected in the vof initial condition with a reinitialization method other than the geometric one. It was because the SignedDistanceSolver was only instantiated if the geometric reinitialization method was selected.

Solution
Add the instantiation of the SignedDistanceSolver if the geometric smoother is selected in the vof initial condition.

Testing
The segmentation fault doesn't happen anymore.

Documentation
No new parameters for now, but the documentation was updated to specify that the initial condition geometric smoother parameters are declares in the reinitialization method subsection.



Co-authored-by: Bruno Blais <blais.bruno@gmail.com>