Add another Clang Tidy action that uses a modern clang version (V20) (#1714)
Description
The Clang Tidy CI uses an older version of clang (V12) and an older version of the ubuntu image (22.04). This PR updates the Clang Tidy CI to use a new version of clang (V20) and a newer version of the ubuntu image (24.04). However, the new Clang Tidy CI identifies a lot of mistakes to fix and this greatly surpasses what can be achieved reasonably in this PR. Consequently, we enable both the V12 and the V20 CI at the same time and until we fix all of the issues of the V20 CI, we will keep both v12 and V20 CI working together.

Testing
The new Clang CI breaks, but this is something we will fix in the future by addressing all of the warnings it identifies.

Documentation
Nothing to describe here.