Remove a substantial amount of CLang Tidy V20 warnings. Mostly in the core library. (#1722)
Description
In our efforts to migrate from Clang Tidy v12 to Clang Tidy v20, this PR does the first batch of refactoring of the core library of lethe. The main things that is changed is that many loops over the dimensions used to use unsigned integer as indices and, when compared to the dimension (dim), this would raise a warning regarding comparison between signed and unsigned integer. This PR fixes that by using integers for these loops.

Other small things are to use enum with a smaller data footprint and some usage of range-based interators. Some spelling mistakes there and there were also fixed. Overall this PR removes half of the CLang V20 warnings.

Testing
None of the tests results change.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>