Refactored code in multiphysics solvers in write_checkpoint and read_… (#1725)
Description
This PR refactors the serialization and deserialization of tables in all classes derived from AuxiliaryPhysics in a similar manner to that of the NavierStokesBase, CFDDEMSovler and luidDynamicsNitsche.

This PR addresses issue #1689. Additionally, a small correction is made in FluidDynamicsNitsche.

Testing
This PR been tested by checkpointing and restarting simulations involving the four Multiphysics solvers. The tables were printed at the restarting point of a simulation, and contained the data from iterations previously run. The output is thus similar to before the modification.

Documentation
No need for additional documentation.