Fix bug with void fraction not being resized correctly (#1732)
Description
The array used to calculate the void fraction at the particle location was not being resized to the number of particles, but it was kept at a constant size corresponding to the maximum number of particles. This could lead to a weird segmentation fault in some edge cases that I have not found an easy way to reproduce.

Solution
This is fixed by resizing the array to be consistent with the number of particles.

Testing
I have not been able to create an application test that reproduces that bug in a reliable fashion, but since this is a minor bug, I don't think it is a major deal.

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>