Write pvd file for boundaries (#1731)
Description
This PR adds a post processing function to write pvd files for boundaries when output boundaries is set to true in the simulation control section. This function is used when the mortar feature is enabled, so that rotating boundaries can be printed at every time step.

Testing
No new tests have been added, but the feature has been tested with a 3D mixer which can de documented as an example in the future.