Migrate Clang TIDY V20 to a custom docker image and self-hosted runners (#1745)
Migrates the CLang V20 Tidy CI to use self-hosted runners. An additional feature is that the runner now uses a custom Docker image that I have made. This is significantly faster than before (10-15 minutes vs 2 hours). It can still be improved, but for now this is much better than what it was before.