dpg electromagnetic solver in parallel (#1712)
Description
This is the current version of the EM-DPG solver in parallel that I used to produce the figures of convergence with Matthias meeting on October 9 2025. Do you think I should change the format to follow the same nomenclature as the other prototype (e.g., ScratchData, CopyData) before merging and starting the implementation in Lethe?

Testing
I ran convergence analysis for DG0, DG1 and DG2.

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>