Add void fraction output at iteration 0 (#1730)
Description
In the CFD-DEM solver, the void fraction was not output at iteration 0. This is a very minor bug and this PR fixes it.

Solution
The output was done in set_initital_condition() before initializing the void fraction. The set_initial_conditions was defined in NavierStokesBase. A new version has been defined in FluidDynamicsVANS that skips the output. The solution is now explicitly written after the void fraction has been initialized.

Testing
This has been tested by launching a simulation and visualizing the void fraction in the vtu files.