Fix compilation adding missing includes (#1750)
Description
In #1740 We replaced some manual sqrt(2) with the use of std::numbers. This is a great idea since this follows the C++20 standard. However, the include was missing in two files, which prevented compilation on Apple CLANG.

Solution
The problem is fixed by adding the correct includes in the two modified files.

Testing
No test output change.