Fix issue with boundary pvd file (#1751)
Description
The writing of a PVD for the boundary (introduced in #1731 ) introduced a bug identified in #1749 that prevents the simulation from adequately writing checkpoint files. This is caused by an incorrect call to read instead of write in the write_checkpoint routine.

Solution
Call the save instead of the read functionality at this location.

Testing
Tested it on my machine and it works just fine. This is a one liner so I don't think this requires a new test.