Coupling modes particle fluid (#1752)
Description
The particle-fluid coupling the vans and cfd-dem solver currently rely on a semi-implicit coupling where the drag force is calculated explicitely and partially implicited on the fluid side. This is a bit problematic for some cases and I want to introduce capability to control what we are essentially doing.

This PR introduces 3 coupling mode. The semi-implicit (which is the one we actually use right now) as well as a fully implicit and a fully explicit coupling mode. In the fully implicit mode, the particle-fluid forces are calculated using the full velocity solution at time t+delta t. Surprisingly, this is a lot more stable than I thought and it leads to very good results. In the fully explicit formulation, the particle-fluid forces are calculated at time t and are applied to both the particle and the fluid. In this case, there is no Jacobian matrix and the solver is fully explicit. This means that these is a stability criterion.

To make this work well, I had to move the CFD-DEM parameters to the core solver so that the library may be used within the scratch data (since I need to know the drag mode in the calculation of the velocity at the particle location). I also found some slight issues with the two-way coupling for the lift forces and there were some signs that were conmfusing (for example the lift force was applied to both the fluid and the particle with the same sign and the - sign came later, this was kinda confusing and I took the opportunity to clarify this here.

Testing
Three tests are added.
explicit-particle-sedimentation -> Sedimentation of a single particle with explicit drag coupling
implicit-particle-sedimentation -> Sedimentation of a single particle with implicit drag coupling
semi-implicit-particle-sedimentation -> Sedimentation of a single particle with semi-implicit drag coupling

They all lead to very similar velocity evolution (implicit and semi-implicit are absurdly close), but explicit has some slight oscillations in the particle velocity which are kinda interesting.

Documentation
The new parameter is documented

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>