Fix interface_tools_sphere_reconstruction.cc warning (#1788)
The warning in tests/core/interface_tools_sphere_reconstruction.cc was because we are creating variable surface_cells by copy and using it as read-only. To fix it, it is created now as a const reference.
The variable surface_cells looks like it is read-only but does not have a similar warning. I made a const reference anyway.