Matrix free implicit coupling (#1757)
Description
This introduces the particle-fluid coupling the VANS matrix-free operator. It was a lot more work than anticipated, but the results are nice.

Essentially, this PR introduces the capability to couple implicitly, explicitly or semi-implicitly the particle-fluid forces onto the fluid. The coupling leverages the QCM projection to project the drag on the particle, the particle velocity and the momentum transfer coefficient onto the CFD mesh. Then, these projected fields are used in the VANS formulation within the coupling.

There is one thing special here. When the semi-implicit and implicit couplings are used, the drag force applied on the fluid is the momentum_transfer_term * (u_particle-u_fluid). This is not numerically fully identical to the filtered drag force. However, this is the only way I have managed to make the coupling robust and stable, so I think this is the way to go forward. Consequently, this PR meant I had to add a new property on the particles (their momentum exchange coefficient), but there is absolutely no way around that right now. The results are very good though for all of the coupling modes and the linear solver is behaving much better than I expected.

Testing
Three new tests have been added that are the particle-fluid coupling with drag explicit, semi-implicit and implicit. These tests perform a simulation through a packed bed and measure the pressure drop through the bed.

Documentation
There is nothing to document for now since this is a very experimental solver.

Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>