Fix bug with particle output (#1800)
Description
A bug had been introduced in [#1752] that prevented paraview from opening the vtu, pvtu and pvd files for the particle results in the CFD-DEM solver. This was because the new property (momentum_transfer_coefficient) was not given a name and this prevented the output from being adequately named.

Solution
This PR fixes this issue by adding the name to the new property. Which in turns fixes the output for the paraview file. It is a one-line change, but that has been consequences.

Testing
Tested by running a few cases on my machine. All is fine. This is relatively hard to test automatically regrettably.