Fix settling cube test case (#1805)
Description
Since PR #1722 , the sedimentation-1-cuboid example crashes when the particle collides with the bottom wall. The previous case used to run, but the non-linear and linear solver struggled pretty hard during the collision. I suspect that the case itself was actually unstable and that PR #1722 did not break anything in particular.

Solution
Using adaptive time-stepping with a max CFL of 0.5 resolves the issue, leads to the same results, and reduces the computational time. We will keep on monitoring this test case to ensure that it remains stable.

Testing
Tested the example with the new parameters on my mac AND on Lucille. It leads to the right results and is robust. It is also much faster to run now (about 2X or so)

Co-authored-by: Olivier Guévremont <guevremont.o@gmail.com>