Remove disabled CI and add prmindent CI (#1808)
Description
The parameter files we use in Lethe are indented following a certain tool that allows the prm files to be easy to read and robust. They can all be indented by running the pre-commit.sh script that is in the contrib/utilities folder. However, we often forget to manually indent the prm files like this.

Consequently, this PR adds a new CI (yet another CI) that checks if the prm indent script would change a prm file. If it does, then the CI fails. If it does not, then it works. This will be a very short CI step (like 10s) so we can run it on the Github Actions runner.

Testing
This version of the branch right now has non-indented prm files, so this CI step should show an error.