Refactor `initial_step` parameter in solve_linear system (#1814)
Description
Issue [#1803] pointed out that the parameter initial_step in solve_linear_system function is not used. This PR is to solve this issue and remove the parameter. In addition, this parameter was coming upstream from solve_non_linear_system so it was removed from there also.

Testing
Lethe test suit passed.

Documentation
The initial_step and first_step bool parameters are not used anymore. The flag was originally used to apply the non_zero_constraints in the different solver at the first iteration, but now this is implemented directly in the linear solver options of physics (i.e., solve_system_GMRES, solve_system_BiCGStab, solve_System, solve_system_direct).

Co-authored-by: Bruno Blais <blais.bruno@gmail.com>