Remove incomplete Reynolds stress solutions from MultiphysicsInterface (#1822)
Description
Reynolds stress solutions were to be integrated into MultiphysicsInterface. The integration work began in 2022 but was never completed. Since they are not used and the implementation that had been started was flawed, the relevant code elements have been removed. This can be reworked when it becomes necessary to integrate Reynolds stress solutions into MultiphysicsInterface.