SDIRK Method with MatrixFree solver (#1616)
Description
This PR enables the use of the SDIRK method with the MatrixFree solver.
It should be seen as a generalization of the previous SDIRK Implementation PR. This new version works for MatrixBased, MatrixFree and Blocks resolution.

The main change is the move of the functions multi_stages_preresolution, multi_stages_postresolution and update_multi_stage_solution. In the last PR, they were implemented only in the NavierStokesBase class. But since we need to use functions located in the FluidDynamicsMatrixFree class inside the function multi_stages_preresolution now, I choosed to use the polymorphism structure and to implement each of these three functions respectively in the FluidDynamicsMatrixBased, FluidDynamicsMatrixFree and FluidDynamicsBlocks classes.

Other changes are minor. Most of them copy the structure of the already existing BDF implementation.

Testing
Two tests have been added. They consist in 2D tgv with two different time-step value. We see that convergence up to 3rd order is reached (in this case it's like 2.87)

Documentation
Now the SDIRK scheme is available for MatrixBased, MatrixFree and Block solvers.

Co-authored-by: etiennemarin <etienne.marin@ensta.fr>
Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>
Co-authored-by: Bruno Blais <blais.bruno@gmail.com>
Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>