Linear solver (#1804)
Description
For the DPG-electromagnetic physics, we don't see the use of a non-linear solver in the near future. Therefore, I modified the current architecture to be able to solve systems in a linear fashion without the need of the non-linear loop. I also created an associated test inspired by the non-linear one.

Testing
The new test added give the right output.

Documentation
The physics_solver and the non_linear_solver classes have now an overloaded constructor to solve linear system. To improve the semantics with this new architecture, the general name "solver_stratgy" is used instead of "non_linear" where approriate. A new linear_solver_strategy class is also introduced with is associated unit test.