Add option to override time step on restart (#1824)
Description
It is often useful to change the time-step before restarting a simulation, but that wasn't possible when using adaptive time-stepping. We add a new parameter to allow this explicitly.

Testing
A new test was added. All previous tests still pass.

Documentation
The new parameter was added to the documentation.

Co-authored-by: Amishga Alphonius <107414376+AmishgaAlphonius@users.noreply.github.com>