Fix rayleigh plateau weber (#1833)
Description
Fix the the Weber number reported in the example. It was stated to be 20 in the .rst, but it is 50.

Miscellaneous (will be removed when merged)
The results reported in the article are for We=50, so we don't need to re-run them now. But at some point we should update the simulation with the newest settings of Lethe.