Fix sign error dissipation heat transfer (#1830)
Description
There was a sign error in the viscous dissipation term of the theory guide. It was written with a - in front of it, but then phi_v was defined as (grad u + grad u^T). This error is most likely coming from using the BSl as a reference material in which the deviatoric stress tensor is - (grad u + grad u ^T) and thus it became (-) * (-) -> +.

The theory guide is now correctly considering the new sign. I also think that using phi as a variable was a bit too vague since we have a gazillion phi. I called it phi_v instead for viscous.