Force Projection in CFD-DEM Matrix Based Solver (#1813)
Description
This PR adds the option of using the filtered particle-fluid forces in the VANS equations for the matrix-based CFD-DEM solver, following the logic adopted by @blaisb . In this PR, only an explicit coupling is implemented for the fluid drag and one test case is performed, which is that of a single sedimenting particle. Subsequent PRs including the implicit and semi-implicit drag couplings, as well as more application tests are to follow.

Testing
This has been tested with the example of one particle that is sedimenting.

Documentation
This PR adds one parameter which is project particle forces. This parameter offers the option to use the projection of particle forces using the QCM filters in the VANS equations.

Also, I modified the python post-processing file to take this into account. Not sure whether it should be included or not.