Particle projector dealii vectors (#1826)
Description
To be compatible with both the matrix-free and the matrix-based architecture, the particle-project needs to support both deal.II parallel vectors and Trilinos parallel vectors. This means that we always have two copies of both dangling around all the time and/or need to convert between them.
One standing issue I have with Lethe is this multiplicity of vector type. As part of the progressive transition towards having only deal.II vectors everywhere instead of the blend of vectors, I wanted to first refactor the ParticleProjector to only use deal.II parallel vectors.

To do this in a step-by-step manner, this PR only refactors the ParticleFields such that they are only deal.II vectors. Consequently, this eliminate a large number of vectors and/or a number of transition stage. To use these vectors in a matrix-based environment is straightforward (they can be used straightup, or they can be converted to a Trilinos vector at the moment of their usage), but at least this is really simpler in terms of the ParticleFields class.

Consequently, the CG solver has changed from the Trilinos one to the deal.II one (not a big change really).
@antonellasuccar this might create a conflict with your open PR. We should check which one to merge first I think.

Testing
I have added a test that tests the particle projector logic in parallel.It is like an existing test, but this one uses two cores instead of one and it also manages a larger mesh. It is better to have a parallel test anyway for these fields to ensure that the ghosting is adequately done.

Documentation
Nothing to document here.
---------

Co-authored-by: Victor Oliveira Ferreira <voferreira@outlook.com>
Co-authored-by: Antonella Succar <142912677+antonellasuccar@users.noreply.github.com>